Success! The installed service starts automatically, and starts Jena Fuseki listening for HTTP on port 4000.
Without the use64BitJre flag set to true, the installed service refused
to start on 64 bit Windows Server 2012.  The service's log, showed
messages like:
 %1 is not a valid Win32 application
but there were no spaces in the path of the executable, which is the
most common cause for this message.

Jena Fuseki uses jetty, and not tomcat, but this stackoverflow answer
gave me the clue to try the flag: http://stackoverflow.com/a/35147880