Now uses log capture to get port number verification in unit test.
The fuseki methods all have void return values, so there is no simple
way to find something to assert on.

Mocked a log4j appender and added it to the root logger, then captured
the logs, found the log message containing the port number and verified
that it had the expected port number.